+++ src/app/collections/inactive/inactive.controller.js
@@ -31,6 +31,7 @@
                     { value: 'Drummond Group', selected: true },
                     { value: 'ICSA Labs', selected: true },
                     { value: 'InfoGard', selected: true },
+                    { value: 'SLI Compliance', selected: true },
                 ],
                 edition: [
                     { value: '2014', selected: true },